拆分 gui_main 系统设置页构建
